requirements.txt
opencv-python~=4.8.1.78
pillow~=10.0.1
torchvision~=0.16.0
numpy~=1.26.0
matplotlib~=3.8.0
tqdm~=4.65.2
scikit-learn~=1.3.0
scipy~=1.11.3
